Switch to GCC 4.2 and quash some warnings (maybe fixing an MSN bug in the process?)

+++ Frameworks/Adium Framework/Source/AIService.h
@@ -59,7 +59,7 @@ typedef enum {
 - (NSCharacterSet *)allowedCharactersForUIDs;
 - (NSCharacterSet *)allowedCharactersForAccountName;
 - (NSCharacterSet *)ignoredCharacters;
-- (int)allowedLength;
+- (NSUInteger)allowedLength;
 - (int)allowedLengthForUIDs;
 - (int)allowedLengthForAccountName;
 - (BOOL)caseSensitive;
