+++ Plugins/Purple Service/ESMSNService.m
@@ -68,7 +68,7 @@
 - (NSCharacterSet *)ignoredCharacters{
 	return [NSCharacterSet characterSetWithCharactersInString:@""];
 }
-- (int)allowedLength{
+- (NSUInteger)allowedLength{
 	return 113;
 }
 - (BOOL)caseSensitive{
